Rename folder vector_quantize_pytorch into vector_quantization
