+++ examples/autoencoder.py
@@ -18,7 +18,7 @@ from torch import nn
 from torch.nn.functional import l1_loss
 
 from examples.data import LitFashionMNIST
-from vector_quantize_pytorch import VectorQuantize
+from vector_quantization import VectorQuantize
 
 seed_everything(1234, workers=True)
