+++ examples/autoencoder_fsq.py
@@ -19,7 +19,7 @@ from torch import nn
 from torch.nn.functional import l1_loss
 
 from examples.data import LitFashionMNIST
-from vector_quantize_pytorch import FSQ
+from vector_quantization import FSQ
 
 seed_everything(1234, workers=True)
