+++ vector_quantization/finite_scalar_quantization.py
