+++ vector_quantization/utils/general.py
