+++ vector_quantization/utils/losses.py
